fixed some visual bugs
