+++ css/menu.css
@@ -1,0 +1,88 @@
+.mobile-menu-hidden {
+	background-color: transparent;
+	z-index: -999;
+}
+
+.mobile-menu-shown {
+	background-color: rgba(0, 0, 0, 0.8);
+	z-index: 999;
+}
+
+.mobile-menu {
+	position: fixed;
+	top: 0;
+	left: 0;
+	width: 100%;
+	height: 100%;
+	display: flex;
+	justify-content: end;
+	align-items: center;
+	backdrop-filter: blur(5px);
+	transition: background-color 0.3s ease-in-out, backdrop-filter 0.3s ease-in-out, z-index 0.3s ease-out;
+}
+
+.mobile-menu-list-hidden {
+	transform: translateX(100%);
+}
+
+.mobile-menu-list-shown {
+	transform: translateX(0);
+}
+
+.mobile-menu-list {
+	list-style: none;
+	height: 100%;
+	background-color: #fff;
+	width: 100%;
+	max-width: 20rem;
+	display: flex;
+	flex-direction: column;
+	justify-content: start;
+	align-items: end;
+	gap: 1rem;
+	padding: 1rem;
+	transition: all 0.3s ease-in-out;
+}
+
+.mobile-menu-list a {
+	color: black;
+}
+
+.mobile-menu-list a:hover {
+	color: #ff4b2b;
+}
+
+.mobile-menu-header {
+	width: 100%;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	gap: 1rem;
+}
+
+.mobile-menu-list .clicker {
+	padding: 0.5rem 1rem;
+	background-color: #ff4b2b;
+	color: white;
+	border-radius: 5px;
+	cursor: pointer
+	border: none;
+	box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
+
+.mobile-menu-header > button {
+	background-color: transparent;
+	border: none;
+	cursor: pointer;
+	padding: 0.5rem 1rem;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+/* End of mobile menu section */
+
+@media only scren and (min-width: 650px) {
+	.mobile-menu {
+		display: none;
+	}
+}
