+++ css/tracking.css
@@ -1,0 +1,102 @@
+#order-status {
+	width: 100%;
+	height: 100%;
+}
+
+#order-status .placeholder {
+	width: 100%;
+	height: 100%;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+
+.order-tracking-background {
+	position: absolute;
+	width: 100%;
+	height: 100%;
+	z-index: 0;
+}
+
+.order-tracking input {
+    background-color: #cfcbcb;
+	padding: 0.5rem 1rem;
+	border: 0px solid #cfcbcb;
+	outline: 0px solid #cfcbcb;
+	border-radius: 5px;
+	transition: outline 0.3s ease-in-out, border 0.3s ease-in-out;
+}
+
+.order-tracking input:focus {
+	border: 1px solid #ff4b2b;
+	outline: 1px solid #ff4b2b;
+}
+
+.order-tracking input {
+	flex: 5;
+}
+
+.order-tracking button {
+	background-color: #ff4b2b;
+	border-radius: 5px;
+	color: white;
+	border: none;
+	padding: 10px;
+	flex-grow: 1;
+	cursor: pointer;
+}
+
+.order-tracking-container-show {
+	opacity: 1;
+	backdrop-filter: blur(5px);
+	z-index: 999;
+}
+
+.order-tracking-container-hide {
+	opacity: 0;
+	backdrop-filter: blur(0px);
+	z-index: -999;
+}
+
+.order-tracking-container {
+	position: fixed;
+	top: 0;
+	left: 0;
+	width: 100%;
+	height: 100%;
+	background-color: rgba(0, 0, 0, 0.5);
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	transition: opacity 0.3s ease-in-out, backdrop-filter 0.3s ease-in-out, z-index 0.3s ease-out;
+}
+
+.order-tracking-show {
+	height: 15rem;
+	padding: 0.5rem 0.5rem;
+}
+
+.order-tracking-hide {
+	height: 0rem;
+	padding: 0rem 0.5rem;
+}
+
+.order-tracking {
+	z-index: 1;
+    background-color: #ffffff;
+    border-radius: 10px;
+    box-shadow: 0 6px 20px rgba(0, 0, 0, 0.3);
+	display: flex;
+	flex-direction: column;
+	gap: 0.5rem;
+	overflow: hidden;
+	transition: height 0.3s ease-in-out;
+}
+
+.order-tracking .input-area {
+	display: flex;
+	justify-content: center;
+	gap: 0.5rem;
+	max-width: 50rem;
+}
+
