+++ js/script.js
@@ -231,18 +231,6 @@ orderStatusBtn.addEventListener("click", () => {
 	})
 })
 
-const headerButtons = document.querySelectorAll(".header-button")
-const mobileMenu = document.querySelector(".mobile-menu")
-const mobileMenuList = document.querySelector(".mobile-menu-list")
-headerButtons.forEach(button => {
-	button.addEventListener("click", () => {
-		mobileMenu.classList.toggle("mobile-menu-hidden")
-		mobileMenu.classList.toggle("mobile-menu-shown")
-		mobileMenuList.classList.toggle("mobile-menu-list-hidden")
-		mobileMenuList.classList.toggle("mobile-menu-list-shown")
-	})
-})
-
 const history = document.querySelector("#history")
 const clicker = document.querySelector(".mobile-menu-header > .clicker")
 const btn = document.createElement("button");
